added directory deletion option

+++ cleanup.py
@@ -3,6 +3,7 @@ from argparse import ArgumentParser
 import re
 import glob
 import os
+import shutil
 import time
 import tarfile
 from datetime import datetime
@@ -28,25 +29,36 @@ def dhms(s):
                 total += int(match.strip(key)) * value
     return total
 
-def cleanup(path, types, retention, action, dryrun=False):
+def cleanup(path, type, globs, retention, action, dryrun=False):
 
     affected_files = 0
 
-    for type in types:
+    for gl in globs:
         
         check_time = time.time() - retention
         now = datetime.now().strftime('%Y-%m-%d-%H-%M-%S')
         
-        g = glob.glob(f'{path}/**/*.{type}', recursive=True)
+        g = glob.glob(f'{path}/**/{gl}', recursive=True)
 
         if action != 'archive' or dryrun:
             for file in g:
-                if os.path.getmtime(file) < check_time:
+
+                if type in ['f', 'file']:
+                    type_check = os.path.isfile(file)
+                elif type in ['d', 'dir', 'directory', 'folder']:
+                    type_check = os.path.isdir(file)
+                else:
+                    type_check = False
+
+                if os.path.exists(file) and type_check and os.path.getmtime(file) < check_time:
                     if action == 'print' or dryrun:
                         logging.info(file)
                     elif action == 'delete':
                         logging.debug(f'deleting {file}')
-                        os.remove(file)
+                        if type in ['f', 'file']:
+                            os.remove(file)
+                        else:
+                            shutil.rmtree(file)
                     affected_files += 1
 
         elif action == 'archive':
@@ -54,14 +66,26 @@ def cleanup(path, types, retention, action, dryrun=False):
             if len(g) > 0:
                 with tarfile.open(output_filename, "w:gz") as tar:
                     for file in g:
-                        if os.path.getmtime(file) < check_time:
+
+                        if type in ['f', 'file']:
+                            type_check = os.path.isfile(file)
+                        elif type in ['d', 'dir', 'directory']:
+                            type_check = os.path.isdir(file)
+                        else:
+                            type_check = False
+
+                        if os.path.exists(file) and type_check and os.path.getmtime(file) < check_time:
                             logging.debug(f'archiving {file} into {output_filename}')
                             tar.add(file, arcname=os.path.basename(file))
                             affected_files += 1
                             
                     for file in g:
-                        logging.debug(f'deleting {file}')
-                        os.remove(file)
+                        if os.path.exists(file) and type_check and os.path.getmtime(file) < check_time:
+                            logging.debug(f'deleting {file}')
+                            if type in ['f', 'file']:
+                                os.remove(file)
+                            else:
+                                shutil.rmtree(file)
     
     return affected_files
     
@@ -93,10 +117,11 @@ def main():
     args = list()
 
     for folder in folders:
-        types = parser[folder]['types'].split(' ')
+        type = parser[folder]['type']
+        globs = parser[folder]['globs'].split(' ')
         retention = dhms(parser[folder]['retention'])
         action = parser[folder]['action']
-        args.append((folder, types, retention, action))
+        args.append((folder, type, globs, retention, action))
 
     affected_files = 0
